Updated code reorganization and adapter ProphetNet
